RFID Bedroom Sensor
The sensor is able to detect collision with the avatar and different
objects or both at the same time. The sensor is divided in 3 sections,
each of them produce a different output value.
For the moment, each sensor is able to store 10 different objects at the
same time (the array is not dynamic).
We need to change shape of the different zones to be more realistic (no
more cubic section)