RFID+ PressurePlate
RFID : Only the children of the gameobject is now disable. It takes 4
second to do a full turn on the token ring. We can adjust that by
changing the value in the InvokeRepeating of the parent class.

PressurePlate : Now uses OnCollisionEnter instead of OnTriggerEnter.
That way, the plate is active when the player is on it and not 1 feet
away like before.